enum changed to enum class

+++ coursework/src/element/ElementCreation.h
@@ -28,7 +28,7 @@ namespace ElementCreation {
         std::vector<SimpleSprite*> sprites;
         
         if (type >= ANIMATED_TYPES_START) {
-            const auto &info = Constants::ANIMATED_SPRITE_INFO[type - ANIMATED_TYPES_START];
+            const auto &info = Constants::ANIMATED_SPRITE_INFO[static_cast<size_t>(type) - static_cast<size_t>(ANIMATED_TYPES_START)];
             for (auto &src : info.first.first) {
                 sprites.push_back(new AnimatedSprite(
                     src + std::to_string(num) + ".png",
@@ -36,7 +36,7 @@ namespace ElementCreation {
             }
         }
         else {
-            const auto &info = Constants::SIMPLE_SPRITE_INFO[type];
+            const auto &info = Constants::SIMPLE_SPRITE_INFO[static_cast<size_t>(type)];
             for (auto &src : info.first)
                 sprites.push_back(new SimpleSprite(src + std::to_string(num) + ".png", info.second));
         }
